portfolio/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
PyYAML==6.0.1
python-multipart==0.0.9
